LayoutTests/imported/w3c/web-platform-tests/webrtc-extensions/RTCRtpParameters-maxFramerate-expected.txt


FAIL addTransceiver() with sendEncoding.maxFramerate field set to less than 0 should reject with RangeError assert_throws_js: function "() => pc.addTransceiver('video', {
    sendEncodings: [{
      maxFramerate: -10
    }]
  })" did not throw
FAIL addTransceiver('audio') with sendEncoding.maxFramerate should succeed, but remove the maxFramerate, even if it is invalid assert_not_own_property: unexpected property "maxFramerate" is found on object
FAIL setParameters with maxFramerate on an audio sender should succeed, but remove the maxFramerate assert_not_own_property: unexpected property "maxFramerate" is found on object
FAIL setParameters with an invalid maxFramerate on an audio sender should succeed, but remove the maxFramerate assert_not_own_property: unexpected property "maxFramerate" is found on object
FAIL setParameters() with encoding.maxFramerate field set to less than 0 should reject with RangeError assert_not_own_property: unexpected property "maxFramerate" is found on object
PASS setParameters() with maxFramerate 24->16 should succeed with RTCRtpTransceiverInit
PASS setParameters() with maxFramerate 24->16 should succeed without RTCRtpTransceiverInit
FAIL setParameters() with maxFramerate undefined->16 should succeed with RTCRtpTransceiverInit assert_equals: expected (undefined) undefined but got (number) 0
FAIL setParameters() with maxFramerate undefined->16 should succeed without RTCRtpTransceiverInit assert_equals: expected (undefined) undefined but got (number) 0
FAIL setParameters() with maxFramerate 24->undefined should succeed with RTCRtpTransceiverInit assert_equals: expected (undefined) undefined but got (number) 24
FAIL setParameters() with maxFramerate 24->undefined should succeed without RTCRtpTransceiverInit assert_equals: expected (undefined) undefined but got (number) 24
PASS setParameters() with maxFramerate 0->16 should succeed with RTCRtpTransceiverInit
PASS setParameters() with maxFramerate 0->16 should succeed without RTCRtpTransceiverInit
FAIL setParameters() with maxFramerate 24->0 should succeed with RTCRtpTransceiverInit assert_equals: expected 0 but got 24
FAIL setParameters() with maxFramerate 24->0 should succeed without RTCRtpTransceiverInit assert_equals: expected 0 but got 24
